WIP: extract-library Phase 3 finish — move proxies to vtsearch/shim
Relocate the dict/list proxy classes (_ProxyDict, _ProxyList) and the
module-level proxy instances (medias, good_votes, bad_votes,
label_history, vote_click_times, vote_region_boxes, last_learned_scores,
textsort_suggestions) out of vtsearch/state/core.py into a new app-side
module: vtsearch/shim/state_proxies.py.

vtsearch/state/__init__.py re-exports the proxies from the shim so
existing app-tier imports (``from vtsearch.state import medias``) keep
working unchanged.  Tests that reach into state.core for the proxies
were updated to import from vtsearch.state or vtsearch.shim.state_proxies.

snapshot_medias / get_media / clear_medias in state/__init__.py now
operate on the active DatasetContext directly instead of going through
the proxy view, since they are app-tier wrappers themselves.

Phase 3 exit criteria met: ``grep -n '^medias\\|^good_votes\\|^_Proxy'``
across the library-candidate paths returns zero hits.  See Phase 3 of
docs/plans/extract-library.md.
