+++ vtsearch/state/core.py
@@ -1,23 +1,25 @@
-"""Core state variables and lock shared by all state submodules.
+"""Core state types and lock shared by all state submodules.
 
-This module defines the mutable global state and the reentrant lock that
-protects it.  All other ``state_*.py`` submodules import variables from
-here rather than defining their own.
+Defines the per-dataset and per-detector context classes
+(:class:`DatasetContext`, :class:`DetectorContext`), the resolution
+functions that find the "active" context for the current request /
+thread, and the reentrant lock that protects all mutable state.
 
 Multi-dataset support
 ---------------------
-Per-dataset state is bundled in ``DatasetContext`` objects.  A context store
-(``_contexts``) maps dataset IDs to their contexts, and ``_active_dataset_id``
-tracks which one the UI is currently interacting with.
+Per-dataset state is bundled in :class:`DatasetContext` objects.  Per-
+detector state lives in :class:`DetectorContext`.  Context stores map
+each ID to its context, and per-request / thread-local resolvers determine
+which one library helpers operate on.
 
-The module-level names ``medias``, ``good_votes``, ``bad_votes``, etc. are
-**proxy objects** that transparently delegate to the active context's
-underlying data structure.  This means all existing code that imports
-these names continues to work without modification — reads and writes go
-through to the active dataset's state.
-
-When no context is active the proxies behave as empty containers (reads
-return nothing, writes are silently discarded or raise where appropriate).
+The app-side facade
+-------------------
+Module-level convenience names (``medias``, ``good_votes``, …) used to
+live here as proxy objects, but they belong to the app layer — the
+library never imports them.  They now live in
+:mod:`vtsearch.shim.state_proxies` and are re-exported from
+:mod:`vtsearch.state` so existing app-tier imports continue to work.
+See Phase 3 of ``docs/plans/extract-library.md``.
 """
 
 from __future__ import annotations
@@ -433,234 +435,13 @@ def clear_all_detector_contexts() -> None:
 
 
 # ---------------------------------------------------------------------------
-# Proxy classes — delegate to the active context's underlying container
-# ---------------------------------------------------------------------------
-
-
-class _ProxyDict(dict):
-    """Dict-like proxy that forwards all operations to a target dict.
-
-    The *target_attr* names the attribute on the active context that holds
-    the real dict.  The *context_fn* is called to find that context.
-    Every method fetches the target lazily so switching contexts is instant.
-
-    Inherits from ``dict`` so that ``isinstance(medias, dict)`` is ``True``
-    and existing code that type-checks works.
-    """
-
-    def __init__(self, target_attr: str, context_fn=None) -> None:
-        # Do NOT call super().__init__() with data — we are a proxy, not a
-        # real container.  The empty super().__init__() satisfies the dict
-        # constructor without storing anything in our own hash table.
-        super().__init__()
-        object.__setattr__(self, "_target_attr", target_attr)
-        object.__setattr__(self, "_context_fn", context_fn or get_active_context)
-
-    def _target(self) -> dict:
-        ctx_fn = object.__getattribute__(self, "_context_fn")
-        return getattr(ctx_fn(), object.__getattribute__(self, "_target_attr"))
-
-    # -- Core dict methods, all forwarded -----------------------------------
-
-    def __getitem__(self, key):
-        return self._target().__getitem__(key)
-
-    def __setitem__(self, key, value):
-        self._target().__setitem__(key, value)
-
-    def __delitem__(self, key):
-        self._target().__delitem__(key)
-
-    def __contains__(self, key):
-        return self._target().__contains__(key)
-
-    def __iter__(self):
-        # Snapshot keys so `for k in proxy` is safe against concurrent
-        # mutation of the underlying dict by another thread.
-        return iter(list(self._target()))
-
-    def __len__(self):
-        return self._target().__len__()
-
-    def __repr__(self):
-        return f"_ProxyDict({object.__getattribute__(self, '_target_attr')!r}, {self._target()!r})"
-
-    def __eq__(self, other):
-        return self._target().__eq__(other)
-
-    def __ne__(self, other):
-        return self._target().__ne__(other)
-
-    def __bool__(self):
-        return bool(self._target())
-
-    def get(self, key, default=None):
-        return self._target().get(key, default)
-
-    def keys(self):
-        return self._target().keys()
-
-    def values(self):
-        return self._target().values()
-
-    def items(self):
-        return self._target().items()
-
-    def pop(self, *args):
-        return self._target().pop(*args)
-
-    def setdefault(self, key, default=None):
-        return self._target().setdefault(key, default)
-
-    def update(self, *args, **kwargs):
-        return self._target().update(*args, **kwargs)
-
-    def clear(self):
-        return self._target().clear()
-
-    def copy(self):
-        return self._target().copy()
-
-    def __or__(self, other):
-        return self._target().__or__(other)
-
-    def __ior__(self, other):
-        target = self._target()
-        target.__ior__(other)
-        return self
-
-    def __reversed__(self):
-        # Snapshot to avoid RuntimeError under concurrent mutation.
-        return reversed(list(self._target()))
-
-
-class _ProxyList(list):
-    """List-like proxy forwarding to the active context's list attribute."""
-
-    def __init__(self, target_attr: str, context_fn=None) -> None:
-        super().__init__()
-        object.__setattr__(self, "_target_attr", target_attr)
-        object.__setattr__(self, "_context_fn", context_fn or get_active_context)
-
-    def _target(self) -> list:
-        ctx_fn = object.__getattribute__(self, "_context_fn")
-        return getattr(ctx_fn(), object.__getattribute__(self, "_target_attr"))
-
-    def __getitem__(self, index):
-        return self._target().__getitem__(index)
-
-    def __setitem__(self, index, value):
-        self._target().__setitem__(index, value)
-
-    def __delitem__(self, index):
-        self._target().__delitem__(index)
-
-    def __contains__(self, item):
-        return self._target().__contains__(item)
-
-    def __iter__(self):
-        # Snapshot elements so iteration is safe against concurrent mutation.
-        return iter(list(self._target()))
-
-    def __len__(self):
-        return self._target().__len__()
-
-    def __repr__(self):
-        return f"_ProxyList({object.__getattribute__(self, '_target_attr')!r}, {self._target()!r})"
-
-    def __eq__(self, other):
-        return self._target().__eq__(other)
-
-    def __bool__(self):
-        return bool(self._target())
-
-    def __add__(self, other):
-        return self._target().__add__(other)
-
-    def __iadd__(self, other):
-        target = self._target()
-        target.__iadd__(other)
-        return self
-
-    def append(self, item):
-        return self._target().append(item)
-
-    def extend(self, items):
-        return self._target().extend(items)
-
-    def insert(self, index, item):
-        return self._target().insert(index, item)
-
-    def remove(self, item):
-        return self._target().remove(item)
-
-    def pop(self, *args):
-        return self._target().pop(*args)
-
-    def clear(self):
-        return self._target().clear()
-
-    def copy(self):
-        return self._target().copy()
-
-    def index(self, value, *args):
-        return self._target().index(value, *args)
-
-    def count(self, value):
-        return self._target().count(value)
-
-    def sort(self, **kwargs):
-        return self._target().sort(**kwargs)
-
-    def reverse(self):
-        return self._target().reverse()
-
-
-# ---------------------------------------------------------------------------
-# Module-level proxy instances — these ARE the public names that all other
-# modules import.  They look and behave like plain dicts/lists but delegate
-# to the appropriate active context.
-#
-# ``medias`` delegates to the active **DatasetContext** (dataset-intrinsic).
-# All vote/label proxies delegate to the active **DetectorContext**.
-# ---------------------------------------------------------------------------
-
-# Clips storage: id -> {id, type, duration, file_size, embedding, media_bytes, media_string, ...}
-medias: dict[int, dict[str, Any]] = _ProxyDict("medias")  # type: ignore[assignment]
-
-# Voting storage (OrderedDict behavior via dict in Python 3.7+)
-good_votes: dict[int, None] = _ProxyDict("good_votes", get_active_detector_context)  # type: ignore[assignment]
-bad_votes: dict[int, None] = _ProxyDict("bad_votes", get_active_detector_context)  # type: ignore[assignment]
-
-# Combined label history: [(media_id, label, timestamp), ...]
-label_history: list[tuple[int, str, float]] = _ProxyList("label_history", get_active_detector_context)  # type: ignore[assignment]
-
-# Click-time tracking: media_id -> click order (1-indexed).
-vote_click_times: dict[int, int] = _ProxyDict("vote_click_times", get_active_detector_context)  # type: ignore[assignment]
-
-# Per-good-vote region boxes: media_id -> (x0, y0, x1, y1) in normalised image
-# coords.  Only populated when the user drew a region as part of a yes-vote;
-# absent for image-level yes-votes and for every no-vote.  Patch-embedder v2.
-vote_region_boxes: dict[int, tuple[float, float, float, float]] = _ProxyDict(  # type: ignore[assignment]
-    "vote_region_boxes",
-    get_active_detector_context,
-)
-
-# Last learned-sort scores: media_id -> score (float in [0, 1]).
-last_learned_scores: dict[int, float] = _ProxyDict("last_learned_scores", get_active_detector_context)  # type: ignore[assignment]
-
-# Text-sort suggestions: text queries that received a Good vote, most recent last.
-textsort_suggestions: list[str] = _ProxyList("textsort_suggestions", get_active_detector_context)  # type: ignore[assignment]
-
-# Find-mode initial labels: media_id -> label assigned by the detector in find-label.
-_find_initial_labels: dict[int, str] = _ProxyDict("find_initial_labels", get_active_detector_context)  # type: ignore[assignment]
-
-
-# ---------------------------------------------------------------------------
 # Scalar state accessors
 # ---------------------------------------------------------------------------
+# These thin helpers wrap "give me the X of whatever context is currently
+# active" so callers that operate on the active context — but don't need
+# a full DatasetContext / DetectorContext reference — can stay one-liners.
 # Dataset-intrinsic scalars (diversity_tree, dataset_display_name) delegate
-# to the active DatasetContext.  Vote-related scalars (click_counter,
+# to the active DatasetContext.  Detector-related scalars (click_counter,
 # inclusion) delegate to the active DetectorContext.
 # ---------------------------------------------------------------------------
